requirements.txt
pandas
numpy
scikit-learn
streamlit
joblib
python-dateutil
matplotlib
